PRSD-NONE: Rename journeyDataExtensions to journeyExtensions; move getLicenceNumberStepIdAndFormModel to PropertyOwnership ext method
